fix memory issue in tests and const prop bugs
